% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_is_unstable}
\alias{bootstrap_is_unstable}
\title{Simple instability flag based on bootstrap SD of criterion importance}
\usage{
bootstrap_is_unstable(boot, thr_sd = 5)
}
\arguments{
\item{boot}{A \code{paprika_bootstrap}.}

\item{thr_sd}{Threshold in percentage points (e.g. 5).}
}
\value{
TRUE if any criterion importance SD exceeds thr_sd.
}
\description{
Simple instability flag based on bootstrap SD of criterion importance
}
