% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{.ror_alt_coef}
\alias{.ror_alt_coef}
\title{Build utility coefficient vector for an alternative}
\usage{
.ror_alt_coef(engine, alt_idx)
}
\description{
Build utility coefficient vector for an alternative
}
\keyword{internal}
