% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{.ror_delta}
\alias{.ror_delta}
\title{Solve min/max utility difference under current constraints (ROR)}
\usage{
.ror_delta(engine, a_idx, b_idx, sense = c("min", "max"))
}
\description{
Solve min/max utility difference under current constraints (ROR)
}
\keyword{internal}
