% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_export.R
\name{engine_audit_export}
\alias{engine_audit_export}
\title{Export audit report to disk}
\usage{
engine_audit_export(engine, path)
}
\arguments{
\item{engine}{A \code{paprika_engine}.}

\item{path}{Path without extension; will write \verb{<path>_questions.csv} and \verb{<path>_run.json}.}
}
\description{
Export audit report to disk
}
