% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{engine_fit_posterior}
\alias{engine_fit_posterior}
\title{Fit Bayesian preference model (logit with monotone increments) via cmdstanr}
\usage{
engine_fit_posterior(
  engine,
  iter = 1000,
  chains = 2,
  seed = NULL,
  backend = c("stan", "approx")
)
}
\arguments{
\item{engine}{A \code{paprika_engine}.}

\item{iter}{Number of iterations.}

\item{chains}{Number of chains.}

\item{seed}{Random seed.}

\item{backend}{Backend to use ("stan" or "approx").}
}
\description{
Fit Bayesian preference model (logit with monotone increments) via cmdstanr
}
