% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{engine_health}
\alias{engine_health}
\title{Health / validation checks for the engine}
\usage{
engine_health(engine)
}
\arguments{
\item{engine}{A \code{paprika_engine} object.}
}
\value{
An object of class \code{paprika_health} with summary metrics.
}
\description{
Health / validation checks for the engine
}
