% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_paprika_selector.R
\name{engine_pick_tradeoff_polytope}
\alias{engine_pick_tradeoff_polytope}
\title{Full-mode tradeoff picker (polytope-sampling + IG + fairness)}
\usage{
engine_pick_tradeoff_polytope(engine, samples = NULL)
}
\description{
Full-mode tradeoff picker (polytope-sampling + IG + fairness)
}
\keyword{internal}
