% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine_reset}
\alias{engine_reset}
\title{Reset engine to initial state (keep domains/settings)}
\usage{
engine_reset(engine)
}
\arguments{
\item{engine}{A \code{paprika_engine} object.}
}
\value{
A reset \code{paprika_engine}.
}
\description{
Reset engine to initial state (keep domains/settings)
}
