% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_paprika_selector.R
\name{engine_should_stop}
\alias{engine_should_stop}
\title{Full-mode stop criterion}
\usage{
engine_should_stop(engine, samples = NULL)
}
\description{
Full-mode stop criterion
}
\keyword{internal}
