% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{necessary_pref}
\alias{necessary_pref}
\title{Necessary / possible preference checks (ROR)}
\usage{
necessary_pref(engine, a_idx, b_idx)
}
\arguments{
\item{engine}{A \code{paprika_engine}.}

\item{a_idx}{Index of alternative A.}

\item{b_idx}{Index of alternative B.}
}
\description{
Necessary / possible preference checks (ROR)
}
