% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{necessary_top3}
\alias{necessary_top3}
\title{Necessary top-3 (pairwise sufficient check)}
\usage{
necessary_top3(engine)
}
\arguments{
\item{engine}{A \code{paprika_engine}.}
}
\description{
Necessary top-3 (pairwise sufficient check)
}
