% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paprika_classic.R
\name{paprika_build_closure}
\alias{paprika_build_closure}
\title{Build transitive closure of preferences}
\usage{
paprika_build_closure(engine)
}
\arguments{
\item{engine}{Paprika engine with decisions}
}
\value{
List with preference graph (adjacency matrix)
}
\description{
Uses decisions to infer implied preferences via transitivity
}
\keyword{internal}
