% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paprika_classic.R
\name{paprika_filter_bank}
\alias{paprika_filter_bank}
\title{Filter PAPRIKA bank by dominance and transitivity}
\usage{
paprika_filter_bank(engine)
}
\arguments{
\item{engine}{Paprika engine}
}
\value{
Updated engine with filtered candidate bank
}
\description{
Removes questions that can be inferred from previous answers
}
\keyword{internal}
