% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paprika_classic.R
\name{paprika_init_classic_bank}
\alias{paprika_init_classic_bank}
\title{Initialize PAPRIKA question bank for classic mode}
\usage{
paprika_init_classic_bank(engine)
}
\arguments{
\item{engine}{Paprika engine}
}
\value{
Updated engine with PAPRIKA candidate bank
}
\description{
Initialize PAPRIKA question bank for classic mode
}
