% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paprika_pareto.R
\name{paprika_pareto_summary}
\alias{paprika_pareto_summary}
\title{Get human-readable Pareto set summary}
\usage{
paprika_pareto_summary(engine)
}
\arguments{
\item{engine}{Paprika engine with pareto set computed}
}
\value{
Character string describing the current Pareto set
}
\description{
Get human-readable Pareto set summary
}
