% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{possible_pref}
\alias{possible_pref}
\title{Possible preference check}
\usage{
possible_pref(engine, a_idx, b_idx)
}
\arguments{
\item{engine}{A \code{paprika_engine}.}

\item{a_idx}{Index of alternative A.}

\item{b_idx}{Index of alternative B.}
}
\description{
Possible preference check
}
