% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{possible_top3}
\alias{possible_top3}
\title{Possible top-3 (pairwise necessary-domination proxy)}
\usage{
possible_top3(engine)
}
\arguments{
\item{engine}{A \code{paprika_engine}.}
}
\description{
Possible top-3 (pairwise necessary-domination proxy)
}
