% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_solution.R
\name{print.engine_solution_check}
\alias{print.engine_solution_check}
\title{Pretty-print an engine solution check}
\usage{
\method{print}{engine_solution_check}(x, ...)
}
\arguments{
\item{x}{A result from \code{\link[=engine_check_solution]{engine_check_solution()}}.}

\item{...}{Ignored.}
}
\description{
Displays summary fields from \code{\link[=engine_check_solution]{engine_check_solution()}}.
}
