% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{print.paprika_engine}
\alias{print.paprika_engine}
\title{Print a PAPRIKA engine summary}
\usage{
\method{print}{paprika_engine}(x, ...)
}
\arguments{
\item{x}{A \code{paprika_engine}.}

\item{...}{Passed to methods.}
}
\description{
Print a PAPRIKA engine summary
}
