% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_export.R
\name{print_clinician_summary}
\alias{print_clinician_summary}
\title{Print clinician summary to console}
\usage{
print_clinician_summary(engine)
}
\arguments{
\item{engine}{A computed engine.}
}
\description{
Print clinician summary to console
}
