% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_literacy.R
\name{print_decision_aid}
\alias{print_decision_aid}
\title{Print visual decision aid to console}
\usage{
print_decision_aid(
  profiles,
  engine,
  top_n = 3,
  context = c("general", "ms", "oncology", "cardiology")
)
}
\arguments{
\item{profiles}{Data frame of treatment profiles.}

\item{engine}{A computed engine.}

\item{top_n}{Number of top options.}

\item{context}{Healthcare context.}
}
\description{
Print visual decision aid to console
}
