% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_export.R
\name{print_patient_summary}
\alias{print_patient_summary}
\title{Print patient summary to console}
\usage{
print_patient_summary(engine)
}
\arguments{
\item{engine}{A computed engine.}
}
\description{
Print patient summary to console
}
