% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{summary.paprika_engine}
\alias{summary.paprika_engine}
\title{Summarize a PAPRIKA engine}
\usage{
\method{summary}{paprika_engine}(object, ...)
}
\arguments{
\item{object}{A \code{paprika_engine}.}

\item{...}{Passed to methods.}
}
\description{
Summarize a PAPRIKA engine
}
